Fix logic & move test
