+++ core/src/test/scala/kafka/utils/LoggingTest.scala
@@ -18,10 +18,12 @@
 package kafka.utils
 
 import java.lang.management.ManagementFactory
-import javax.management.ObjectName
 
-import org.junit.Test
+import com.typesafe.scalalogging.Logger
+import javax.management.ObjectName
+import org.junit.{Assert, Test}
 import org.junit.Assert.{assertEquals, assertTrue}
+import org.slf4j.LoggerFactory
 
 
 class LoggingTest extends Logging {
@@ -65,5 +67,21 @@ class LoggingTest extends Logging {
     val logging = new TestLogging
 
     assertEquals(logging.getClass.getName, logging.log.underlying.getName)
+  }
+
+  @Test
+  def testLoggerLevelIsResolved(): Unit = {
+    val controller = new Log4jController()
+    val previousLevel = controller.getLogLevel("kafka")
+    try {
+      controller.setLogLevel("kafka", "TRACE")
+      Logger(LoggerFactory.getLogger("kafka.utils.Log4jControllerTest")).trace("test")
+      Assert.assertEquals("TRACE", controller.getLogLevel("kafka"))
+      Assert.assertEquals("TRACE", controller.getLogLevel("kafka.utils.Log4jControllerTest"))
+      Assert.assertTrue(controller.getLoggers.contains("kafka=TRACE"))
+      Assert.assertTrue(controller.getLoggers.contains("kafka.utils.Log4jControllerTest=TRACE"))
+    } finally {
+      controller.setLogLevel("kafka", previousLevel)
+    }
   }
 }
